pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-tagging-service"
version = "0.1.0"
description = "AI-powered image classification and hierarchical tag suggestion service using Gemma 4"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
    "transformers>=4.45",
    "torch>=2.0",
    "accelerate>=0.30",
    "rapidfuzz>=3.0",
    "structlog>=24.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.6",
]

[project.scripts]
image-tagging-service = "image_tagging_service.main:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "PT"]
ignore = ["S101", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
